+++ lib/src/languages/wasm/wasm.dart
@@ -48,6 +48,8 @@ class Wasm {
 
   static const unreachable = 0x00;
 
+  static List<int> br(int i) => <int>[0x0c, ...Leb128.encodeUnsigned(i)];
+
   static List<int> brIf(int i) => <int>[0x0d, ...Leb128.encodeUnsigned(i)];
 
   static List<int> call(int i) => <int>[0x10, ...Leb128.encodeUnsigned(i)];